requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0

# Data labeling and image processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
websockets>=11.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.0.0
pylint>=3.0.0
mypy>=1.0.0
pre-commit>=3.0.0
